requirements.txt
streamlit
python-docx
fpdf2
openai
